# apps/api/openapi.json
{
  "openapi": "3.2.0",
  "paths": {
    "/api/user": {
      "get": {
        "operationId": "me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      },
      "patch": {
        "operationId": "update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/u"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/s"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "operationId": "refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/a"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "logout",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/register": {
      "post": {
        "operationId": "register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/u"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/entries": {
      "get": {
        "operationId": "retrieveEntries",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Entries"
        ]
      },
      "post": {
        "operationId": "createEntry",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/u"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Entries"
        ]
      }
    },
    "/api/entries/sync": {
      "post": {
        "operationId": "syncEntries",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Array of entries to sync from client",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/u"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Entries"
        ]
      }
    },
    "/api/entries/{id}": {
      "delete": {
        "operationId": "deleteEntry",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Entries"
        ]
      },
      "patch": {
        "operationId": "updateEntry",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/c"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Entries"
        ]
      }
    },
    "/api/health": {
      "get": {
        "operationId": "getHealth",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Health"
        ]
      }
    },
    "/api/releases/check": {
      "get": {
        "operationId": "checkUpdate",
        "parameters": [
          {
            "name": "version",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Releases"
        ]
      }
    }
  },
  "info": {
    "title": "JW Service Tracker API",
    "description": "API documentation for the JW Service Tracker application",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [
    {
      "url": "https://jw-service-tracker.vercel.app",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Bearer token authentication"
      }
    },
    "schemas": {
      "u": {
        "type": "object",
        "properties": {
          "preachingDate": {
            "type": "number",
            "description": "Preaching date as Unix timestamp (milliseconds)",
            "example": 1719446400000
          },
          "hours": {
            "type": "integer",
            "description": "Hours spent in preaching activity (0-24)",
            "example": 2,
            "minimum": 0,
            "maximum": 24
          },
          "minutes": {
            "type": "integer",
            "description": "Minutes spent in preaching activity (0-59)",
            "example": 30,
            "minimum": 0,
            "maximum": 59
          },
          "type": {
            "type": "string",
            "description": "Type of preaching session",
            "enum": [
              "house_to_house",
              "revisits",
              "bible_study",
              "other"
            ],
            "example": "house_to_house"
          },
          "notes": {
            "type": "string",
            "description": "Optional notes about the session",
            "example": "Good conversation",
            "maxLength": 50
          },
          "tempId": {
            "type": "string",
            "description": "Temporary ID for offline sync",
            "example": "temp_123abc"
          }
        },
        "required": [
          "preachingDate",
          "hours",
          "minutes",
          "type"
        ]
      },
      "s": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "description": "Phone number with country code",
            "example": "+51912345678"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "password123"
          },
          "platform": {
            "type": "string",
            "description": "Platform identifier",
            "example": "web"
          },
          "deviceName": {
            "type": "string",
            "description": "Device name for session tracking",
            "example": "iPhone 12"
          }
        },
        "required": [
          "phone",
          "password"
        ]
      },
      "a": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Refresh token for obtaining a new access token",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "c": {
        "type": "object",
        "properties": {
          "preachingDate": {
            "type": "number",
            "description": "Preaching date as Unix timestamp (milliseconds)",
            "example": 1719446400000
          },
          "hours": {
            "type": "integer",
            "description": "Hours spent in preaching activity (0-24)",
            "example": 2,
            "minimum": 0,
            "maximum": 24
          },
          "minutes": {
            "type": "integer",
            "description": "Minutes spent in preaching activity (0-59)",
            "example": 30,
            "minimum": 0,
            "maximum": 59
          },
          "type": {
            "type": "string",
            "description": "Type of preaching session",
            "enum": [
              "house_to_house",
              "revisits",
              "bible_study",
              "other"
            ],
            "example": "house_to_house"
          },
          "notes": {
            "type": "string",
            "description": "Optional notes about the session",
            "example": "Good conversation",
            "maxLength": 50
          },
          "tempId": {
            "type": "string",
            "description": "Temporary ID for offline sync",
            "example": "temp_123abc"
          },
          "id": {
            "type": "string",
            "description": "Unique identifier of the entry",
            "example": "507f1f77bcf86cd799439011"
          }
        },
        "required": [
          "preachingDate",
          "hours",
          "minutes",
          "type",
          "id"
        ]
      }
    }
  }
}